Main Menu
Redesigned the main menu. Now the menu for any language works with one function.
Fixed bugs in pictFinder.
The first launch of the bot has been rethought. Now, at the first launch, a request is received to enter a token, after which it is written to settings.txt (created automatically at startup)